remove unused nvidia package
