--- pkg/plugin/nvidia/const.go
@@ -1,43 +1,0 @@
-/*
-Copyright 2023 The Volcano Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package nvidia
-
-const (
-
-	// Extended gpu resource
-	VolcanoGPUMemory = "volcano.sh/vgpu-memory"
-	// Virtual GPU card number
-	VolcanoGPUNumber = "volcano.sh/gpu-number"
-
-	// Annotation
-	PredicateTime = "volcano.sh/predicate-time"
-	GPUAssigned   = "volcano.sh/gpu-assigned"
-	GPUIndex      = "volcano.sh/gpu-index"
-
-	// UnhealthyGPUIDs list of unhealthy gpu ids
-	UnhealthyGPUIDs = "volcano.sh/gpu-unhealthy-ids"
-
-	// Container env
-	// Allocated gpu memory
-	AllocatedGPUResource = "VOLCANO_GPU_ALLOCATED"
-	// Total gpu mem
-	TotalGPUMemory = "VOLCANO_GPU_MEMORY_TOTAL"
-	// NVIDIA visible devices
-	VisibleDevice = "NVIDIA_VISIBLE_DEVICES"
-
-	deviceName = "gpu"
-)
